tailwind: Webinar main page
